+++ src/Graphics/ITexture.hpp
@@ -18,7 +18,7 @@ namespace anvil
     class ITexture : public IResource
     {
     public:
-		virtual void CreateArray(int size, int levels, int width, int height, const gli::format format, const gli::swizzles swizzles) = 0;
+		virtual void CreateArray(size_t size, size_t levels, int width, int height, const gli::format format, const gli::swizzles swizzles) = 0;
 		virtual bool SetLevel(int level, const gli::texture& tex) = 0;
         virtual bool Load(const gli::texture& tex) = 0;
         virtual void Bind() = 0;
